bug 272398 - improved error message

+++ org.eclipse.pde.build/src/org/eclipse/pde/internal/build/BrandingIron.java
@@ -233,13 +233,17 @@ public class BrandingIron implements IXMLConstants {
 			templateLauncher = new File(root, "launcher.exe"); //$NON-NLS-1$
 		if (!templateLauncher.exists())
 			templateLauncher = new File(root, "eclipse.exe"); //$NON-NLS-1$
-		if (brandIcons && templateLauncher.exists()) {
-			String[] args = new String[icons.length + 1];
-			args[0] = templateLauncher.getAbsolutePath();
-			System.arraycopy(icons, 0, args, 1, icons.length);
-			IconExe.main(args);
+		if (brandIcons) {
+			if (templateLauncher.exists()) {
+				String[] args = new String[icons.length + 1];
+				args[0] = templateLauncher.getAbsolutePath();
+				System.arraycopy(icons, 0, args, 1, icons.length);
+				IconExe.main(args);
+			} else {
+				System.out.println("Warning: Could not find executable to brand."); //$NON-NLS-1$
+			}
 		}
-		if (!templateLauncher.getName().equals(name + ".exe")) //$NON-NLS-1$
+		if (templateLauncher.exists() && !templateLauncher.getName().equals(name + ".exe")) //$NON-NLS-1$
 			templateLauncher.renameTo(new File(root, name + ".exe")); //$NON-NLS-1$
 	}
